chat_app/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.1.3
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-countries==7.6.1
django-filter==24.2
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
django-ses==4.1.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
google-auth==2.32.0
gunicorn==20.1.0
h11==0.14.0
idna==3.7
jmespath==1.0.1
jwcrypto==1.5.6
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
paramiko==3.4.0
pillow==10.3.0
psycopg2==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
PyNaCl==1.5.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytweening==1.2.0
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
s3transfer==0.10.2
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
storage==0.0.4.3
sympy==1.13.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
websocket-client==1.8.0
wsproto==1.2.0
